+++ lib/ts/recipe/emailpassword/components/themes/resetPasswordUsingToken/resetPasswordEmail.tsx
@@ -23,6 +23,7 @@ import { useTranslation } from "../../../../../translation/translationContext";
 import { validateForm } from "../../../../../utils";
 import GeneralError from "../../library/generalError";
 import { useUserContext } from "../../../../../usercontext";
+import BackToSignInButton from "../../library/backToSignInButton";
 import BackButton from "../../library/backButton";
 
 const EmailPasswordResetPasswordEmail: React.FC<EnterEmailProps> = (props) => {
@@ -30,6 +31,7 @@ const EmailPasswordResetPasswordEmail: React.FC<EnterEmailProps> = (props) => {
     const t = useTranslation();
     const userContext = useUserContext();
     const [status, setStatus] = useState<EnterEmailStatus>("READY");
+    const [emailFieldValue, setEmailFieldValue] = useState<string>("");
 
     const onSuccess = (): void => {
         setStatus("SENT");
@@ -40,6 +42,13 @@ const EmailPasswordResetPasswordEmail: React.FC<EnterEmailProps> = (props) => {
     };
     const { formFields } = props;
 
+    const emailSuccessText =
+        t("EMAIL_PASSWORD_RESET_SEND_BEFORE_EMAIL") +
+        (emailFieldValue !== undefined && emailFieldValue.length > 0
+            ? emailFieldValue
+            : t("EMAIL_PASSWORD_RESET_SEND_FALLBACK_EMAIL")) +
+        t("EMAIL_PASSWORD_RESET_SEND_AFTER_EMAIL");
+
     if (status === "SENT") {
         return (
             <div data-supertokens="container" css={styles.container}>
@@ -47,11 +56,15 @@ const EmailPasswordResetPasswordEmail: React.FC<EnterEmailProps> = (props) => {
                     <div
                         data-supertokens="primaryText enterEmailSuccessMessage"
                         css={[styles.primaryText, styles.enterEmailSuccessMessage]}>
-                        {t("EMAIL_PASSWORD_RESET_SEND_SUCCESS")}
-                        <span data-supertokens="link" css={styles.link} onClick={resend}>
+                        {emailSuccessText}
+                        <span
+                            data-supertokens="link resendEmailLink"
+                            css={[styles.link, styles.resendEmailLink]}
+                            onClick={resend}>
                             {t("EMAIL_PASSWORD_RESET_RESEND_LINK")}
                         </span>
                     </div>
+                    <BackToSignInButton onClick={props.onBackButtonClicked} />
                 </div>
             </div>
         );
@@ -95,6 +108,14 @@ const EmailPasswordResetPasswordEmail: React.FC<EnterEmailProps> = (props) => {
                                 status: "FIELD_ERROR",
                                 formFields: validationErrors,
                             };
+                        }
+
+                        const emailField = formFields.find((field) => {
+                            return field.id === "email";
+                        });
+
+                        if (emailField !== undefined) {
+                            setEmailFieldValue(emailField.value);
                         }
 
                         return await props.recipeImplementation.sendPasswordResetEmail({