+++ lib/ts/recipe/passwordless/components/themes/styles.ts
@@ -52,18 +52,6 @@ export function getStyles(palette: NormalisedPalette): NormalisedDefaultStyles {
                 textOverflow: "ellipsis",
             },
         },
-        secondaryLinkWithLeftArrow: {
-            marginTop: "10px",
-            marginBottom: "40px",
-            cursor: "pointer",
-            "& svg": {
-                marginRight: "0.3em",
-            },
-            "&:hover svg": {
-                position: "relative",
-                left: "-4px",
-            },
-        },
         resendCodeBtn: {
             width: "auto",
             marginTop: 0,